Work on the billboarding oof the markers.
OpenGL is really pissing me off here. It seems that objects with
alpha transparency need to be depth sorted and drawn in negative depth order - however, the problem is that the point array needs to be sorted in screen coordinates (i.e. after the pushMatrix has been applied) but the points themselves are drawn in model coordinates (BEFORE the rotation is applied) it means that everything needs to be done twice, and probably quite inefficiently.

It's definitely doable, just needs some matrix thought put into it.

Looking good though, nearly there, I'd say.